server/requirements.txt
Flask==3.0.0
aioquic==0.9.23
gunicorn==21.2.0
google-cloud-logging==3.8.0
